requirements.txt
Flask
sklearn-pandas
pandas
numpy
matplotlib
seaborn
scikit-learn